+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CMakeLists.txt
 - README.md CI badge
 - Base (real base)
+- Basic architecture of library : Shape, Point, Drawable, ICanvas and Scene to draw an image
 
 [unreleased]: https://github.com/DostLeFan/ImageDraw/compare/v0.0.0...HEAD
 [v0.0.0]: https://github.com/DostLeFan/ImageDraw/releases/tag/v0.0.0